+++ package.json
@@ -59,7 +59,7 @@
     "@mui/icons-material": "^5.8.4",
     "@mui/lab": "^5.0.0-alpha.108",
     "@mui/material": "^5.11.15",
-    "@rainbow-me/rainbowkit": "^1.1.3",
+    "@rainbow-me/rainbowkit": "^1.2.0",
     "@sentry/react": "^7.43.0",
     "@sentry/tracing": "^7.43.0",
     "@types/jest": "^27.5.2",
